+++ CHANGES.rst
@@ -14,6 +14,9 @@ Version 3.x.x
 - Do not render a ``for`` attribute on the main label of :class:`~fields.RadioField`.
   :issue:`775`
 - Deprecate :class:`~fields.DateTimeField`, which will be removed in WTForms 3.4. :issue:`831`
+- Accept :class:`uuid.UUID` instances in the :class:`~validators.UUID` validator,
+  and translate unexpected types into :class:`~validators.ValidationError`
+  instead of letting :class:`TypeError` escape. :issue:`549` :pr:`769`
 
 Version 3.2.2
 -------------
@@ -100,7 +103,6 @@ Released 2023-10-10
 -   Python 3.11 support :pr:`763`
 -   Added shorter format to :class:`~fields.DateTimeLocalField`
     defaults :pr:`761`
--   Fixed UUID validation :pr:`769`
 -   Stop support for python 3.7 :pr:`794`
 -   Added shorter format to :class:`~fields.WeekField`
     defaults :pr:`765`